move db logic to the manager.py under a challenge manager
