+++ bot/bot.py
@@ -16,7 +16,8 @@ from decouple import config
 from bot.api_client import TortoiseAPI
 from bot.constants import bot_log_channel_id, github_repo_link
 from bot.manager import (
-    Database, ProgressionManager, AFKManager, PointsManager, RetentionManager, TeamManager, GiveawayManager, DutyManager
+    Database, ProgressionManager, AFKManager, PointsManager, ChallengeManager,
+    RetentionManager, TeamManager, GiveawayManager, DutyManager
 )
 from bot.utils.embed_handler import simple_embed
 from bot.utils.error_handler import TortoiseCommandTree
@@ -76,6 +77,7 @@ class Bot(commands.Bot):
         self.db = None
         self.progression_manager = None
         self.points_manager = None
+        self.challenge_manager = None
         self.afk_manager = None
         self.retention_manager = None
         self.team_manager = None
@@ -175,6 +177,7 @@ class Bot(commands.Bot):
             self.progression_manager = ProgressionManager(self.db)
             self.afk_manager = AFKManager(self.db)
             self.points_manager = PointsManager(self.db)
+            self.challenge_manager = ChallengeManager(self.db)
             self.retention_manager = RetentionManager(self.db)
             self.team_manager = TeamManager(self.db)
             self.giveaway_manager = GiveawayManager(self.db)
@@ -183,6 +186,7 @@ class Bot(commands.Bot):
             await self.progression_manager.setup()
             await self.afk_manager.setup()
             await self.points_manager.setup()
+            await self.challenge_manager.setup()
             await self.retention_manager.setup()
             await self.team_manager.setup()
             await self.giveaway_manager.setup()
